Move methods from helper class to upper class.
